add ( home ) : add WhatsAppBubbleDrag
